+++ sceptre/helpers.py
@@ -11,6 +11,8 @@ import sys
 from concurrent.futures import ThreadPoolExecutor
 from concurrent.futures import as_completed
 
+from .exceptions import CircularDependenciesError
+
 
 def camel_to_snake_case(string):
     """
@@ -176,3 +178,36 @@ def get_subclasses(class_type, directory=None):
                         classes[camel_to_snake_case(attr.__name__)] = attr
 
     return classes
+
+
+def _detect_cycles(node, encountered_nodes, available_nodes, path):
+    """
+    Use Depth-first search to detect cycles.
+    :returns: A dictionary containing all of the nodes encountered
+    during the depth first search.
+    """
+    for dependency_name in node.dependencies:
+        # The keys in _load_nodes() (where the available_nodes comes from)
+        # are prefixed with environment names separated by /, which is
+        # undesirable here, so we strip them out.
+        dependency = available_nodes[dependency_name.split("/")[-1]]
+        status = encountered_nodes.get(dependency)
+        if status == "ENCOUNTERED":
+            # Reformat path to only include the cycle
+            path.append(dependency_name)
+            cycle = path[path.index(dependency_name):]
+            raise CircularDependenciesError(
+                "Found circular dependency involving "
+                "{0}".format(cycle)
+            )
+        elif status is None:
+            encountered_nodes[dependency] = "ENCOUNTERED"
+            path.append(dependency_name)
+            _detect_cycles(
+                dependency,
+                encountered_nodes,
+                available_nodes,
+                path
+            )
+            encountered_nodes[dependency] = "DONE"
+    return encountered_nodes