+++ tests/test_environment.py
@@ -2,7 +2,7 @@
 
 import os
 import pytest
-from mock import patch, sentinel, Mock, PropertyMock
+from mock import patch, sentinel, MagicMock, Mock, PropertyMock
 
 from botocore.exceptions import ClientError
 
@@ -12,6 +12,7 @@ from sceptre.exceptions import InvalidEnvironmentPathError
 
 from sceptre.environment import Environment
 from sceptre.stack_status import StackStatus
+from sceptre.stack import Stack
 
 
 class TestEnvironment(object):
@@ -148,9 +149,7 @@ class TestEnvironment(object):
 
         self.environment.launch()
 
-        mock_check_for_circular_dependencies.assert_called_once_with(
-            sentinel.dependencies
-        )
+        mock_check_for_circular_dependencies.assert_called_once_with()
         mock_build.assert_called_once_with(
             "launch", sentinel.threading_events,
             sentinel.stack_statuses, sentinel.dependencies
@@ -178,9 +177,7 @@ class TestEnvironment(object):
 
         self.environment.delete()
 
-        mock_check_for_circular_dependencies.assert_called_once_with(
-            sentinel.dependencies
-        )
+        mock_check_for_circular_dependencies.assert_called_once_with()
         mock_build.assert_called_once_with(
             "delete", sentinel.threading_events,
             sentinel.stack_statuses, sentinel.dependencies
@@ -403,22 +400,221 @@ class TestEnvironment(object):
         }
 
     def test_check_for_circular_dependencies_with_circular_dependencies(self):
-        dependencies = {
-            "stack-1": ["stack-2"],
-            "stack-2": ["stack-1"]
+        stack1 = MagicMock(Spec=Stack)
+        stack2 = MagicMock(Spec=Stack)
+        stack1.dependencies = ["stack2"]
+        stack1.name = "stack1"
+        stack2.dependencies = ["stack1"]
+        stack2.name = "stack2"
+        stacks = {
+            "stack1": stack1,
+            "stack2": stack2
         }
+        self.environment.stacks = stacks
+        with pytest.raises(CircularDependenciesError) as ex:
+            self.environment._check_for_circular_dependencies()
+        assert all(x in str(ex) for x in ['stack2', 'stack1'])
 
-        with pytest.raises(CircularDependenciesError):
-            self.environment._check_for_circular_dependencies(dependencies)
-
-    def test_check_for_circular_dependencies_without_find_dependencies(self):
-        dependencies = {
-            "stack-1": ["stack-2"],
-            "stack-2": []
+    def test_circular_dependencies_with_3_circular_dependencies(self):
+        stack1 = MagicMock(Spec=Stack)
+        stack2 = MagicMock(Spec=Stack)
+        stack3 = MagicMock(Spec=Stack)
+        stack1.dependencies = ["stack2"]
+        stack1.name = "stack1"
+        stack2.dependencies = ["stack3"]
+        stack2.name = "stack2"
+        stack3.dependencies = ["stack1"]
+        stack3.name = "stack3"
+        stacks = {
+            "stack1": stack1,
+            "stack2": stack2,
+            "stack3": stack3
         }
+        self.environment.stacks = stacks
+        with pytest.raises(CircularDependenciesError) as ex:
+            self.environment._check_for_circular_dependencies()
+        assert all(x in str(ex) for x in ['stack3', 'stack2', 'stack1'])
 
+    def test_no_circular_dependencies_throws_no_error(self):
+        stack1 = MagicMock(Spec=Stack)
+        stack2 = MagicMock(Spec=Stack)
+        stack1.dependencies = ["stack2"]
+        stack1.name = "stack1"
+        stack2.dependencies = []
+        stack2.name = "stack2"
+        stacks = {
+            "stack1": stack1,
+            "stack2": stack2
+        }
+        self.environment.stacks = stacks
         # Check this runs without throwing an exception
-        self.environment._check_for_circular_dependencies(dependencies)
+        self.environment._check_for_circular_dependencies()
+
+    def test_DAG_diamond_throws_no_circ_dependencies_error(self):
+        """
+        Ensures
+            o
+           / \
+          o   o
+           \ /
+            o
+        throws no circular dependency error
+        """
+        stack1 = MagicMock(Spec=Stack)
+        stack2 = MagicMock(Spec=Stack)
+        stack3 = MagicMock(Spec=Stack)
+        stack4 = MagicMock(Spec=Stack)
+        stack1.dependencies = ["stack2", "stack3"]
+        stack1.name = "stack1"
+        stack2.dependencies = ["stack4"]
+        stack2.name = "stack2"
+        stack3.dependencies = ["stack4"]
+        stack3.name = "stack3"
+        stack4.dependencies = []
+        stack4.name = "stack4"
+        stacks = {
+            "stack1": stack1,
+            "stack2": stack2,
+            "stack3": stack3,
+            "stack4": stack4
+        }
+        self.environment.stacks = stacks
+        self.environment._check_for_circular_dependencies()
+
+    def test_modified_DAG_diamond_throws_no_circ_dependencies_error(self):
+        """
+        Ensures
+            o
+           / \
+          o   o
+           \ / \
+            o   o
+        throws no circular dependency error
+        """
+        stack1 = MagicMock(Spec=Stack)
+        stack2 = MagicMock(Spec=Stack)
+        stack3 = MagicMock(Spec=Stack)
+        stack4 = MagicMock(Spec=Stack)
+        stack5 = MagicMock(Spec=Stack)
+        stack1.dependencies = ["stack2", "stack3"]
+        stack1.name = "stack1"
+        stack2.dependencies = ["stack4"]
+        stack2.name = "stack2"
+        stack3.dependencies = ["stack4", "stack5"]
+        stack3.name = "stack3"
+        stack4.dependencies = []
+        stack4.name = "stack4"
+        stack5.dependencies = []
+        stack5.name = "stack5"
+        stacks = {
+            "stack1": stack1,
+            "stack2": stack2,
+            "stack3": stack3,
+            "stack4": stack4,
+            "stack5": stack5
+        }
+        self.environment.stacks = stacks
+        self.environment._check_for_circular_dependencies()
+
+    def test_DAG_diamond_with_triangle_throws_no_circ_dependencies_error(self):
+        """
+        Ensures
+            o
+           / \
+          o   o
+           \ / \
+            o ->o
+        throws no circular dependency error
+        """
+        stack1 = MagicMock(Spec=Stack)
+        stack2 = MagicMock(Spec=Stack)
+        stack3 = MagicMock(Spec=Stack)
+        stack4 = MagicMock(Spec=Stack)
+        stack5 = MagicMock(Spec=Stack)
+        stack1.dependencies = ["stack2", "stack3"]
+        stack1.name = "stack1"
+        stack2.dependencies = ["stack4"]
+        stack2.name = "stack2"
+        stack3.dependencies = ["stack4", "stack5"]
+        stack3.name = "stack3"
+        stack4.dependencies = ["stack5"]
+        stack4.name = "stack4"
+        stack5.dependencies = []
+        stack5.name = "stack5"
+        stacks = {
+            "stack1": stack1,
+            "stack2": stack2,
+            "stack3": stack3,
+            "stack4": stack4,
+            "stack5": stack5
+        }
+        self.environment.stacks = stacks
+        self.environment._check_for_circular_dependencies()
+
+    def test_4_cycle_throws_circ_dependencies_error(self):
+        """
+        Ensures
+            o - o
+            |   |
+            o - o
+        throws circular dependency error
+        """
+        stack1 = MagicMock(Spec=Stack)
+        stack2 = MagicMock(Spec=Stack)
+        stack3 = MagicMock(Spec=Stack)
+        stack4 = MagicMock(Spec=Stack)
+        stack1.dependencies = ["stack4"]
+        stack1.name = "stack1"
+        stack2.dependencies = ["stack1"]
+        stack2.name = "stack2"
+        stack3.dependencies = ["stack2"]
+        stack3.name = "stack3"
+        stack4.dependencies = ["stack3"]
+        stack4.name = "stack4"
+        stacks = {
+            "stack1": stack1,
+            "stack2": stack2,
+            "stack3": stack3,
+            "stack4": stack4
+        }
+        self.environment.stacks = stacks
+        with pytest.raises(CircularDependenciesError) as ex:
+            self.environment._check_for_circular_dependencies()
+        assert all(x in str(ex) for x in ['stack4', 'stack3', 'stack2',
+                                          'stack1'])
+
+    def test_modified_3_cycle_throws_circ_dependencies_error(self):
+        """
+        Ensures
+            o   o
+             \ / \
+              o - o
+              (right triangle is a 3 cycle)
+        throws circular dependency error
+        """
+        stack1 = MagicMock(Spec=Stack)
+        stack2 = MagicMock(Spec=Stack)
+        stack3 = MagicMock(Spec=Stack)
+        stack4 = MagicMock(Spec=Stack)
+        stack1.dependencies = ["stack2"]
+        stack1.name = "stack1"
+        stack2.dependencies = ["stack3"]
+        stack2.name = "stack2"
+        stack3.dependencies = ["stack4"]
+        stack3.name = "stack3"
+        stack4.dependencies = ["stack2"]
+        stack4.name = "stack4"
+        stacks = {
+            "stack1": stack1,
+            "stack2": stack2,
+            "stack3": stack3,
+            "stack4": stack4
+        }
+        self.environment.stacks = stacks
+        with pytest.raises(CircularDependenciesError) as ex:
+            self.environment._check_for_circular_dependencies()
+        assert (all(x in str(ex) for x in ['stack4', 'stack3', 'stack2']) and
+                'stack1' not in str(ex))
 
     @patch("sceptre.environment.Config")
     def test_get_config(self, mock_Config):